+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Next Release
 * Feature: Blueprint router can now take a blueprint instance
 * Feature: Added support for ``VIA_ROUTES_NAME`` to set a common routes name
 * Deprecated: ``Basic`` Router in favour of the ``Functional`` router
+* Improved: ``Pluggable`` Router API is now cleaner
 * Improved: Test Suite now uses PyTest
 * Improved: ``ImproperlyConfigured`` now raised if routes module is not defined
   in either ``init_app`` or in application configuration via